TW-1809: refactor
